requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
django-faker==0.2
django-guardian==1.4.9
fake-factory==9999.9.9
gunicorn==19.7.1
idna==2.6
pytz==2017.2
requests==2.18.4
selenium==3.6.0
six==1.11.0
urllib3==1.22
